test: add e2e tests for `getFreePort` api method
