+++ test/server/Server.test.js
@@ -1,7 +1,6 @@
 "use strict";
 
 const { relative, sep } = require("path");
-const http = require("http");
 const webpack = require("webpack");
 const sockjs = require("sockjs/lib/transport");
 const Server = require("../../lib/Server");
@@ -240,144 +239,6 @@ describe("Server", () => {
       require("../../lib/Server");
 
       expect(process.env.WEBPACK_SERVE).toBe(true);
-    });
-  });
-
-  describe("Server.getFreePort", () => {
-    let dummyServers = [];
-
-    afterEach(() => {
-      delete process.env.WEBPACK_DEV_SERVER_BASE_PORT;
-      delete process.env.WEBPACK_DEV_SERVER_PORT_RETRY;
-
-      return dummyServers
-        .reduce(
-          (p, server) =>
-            p.then(
-              () =>
-                new Promise((resolve) => {
-                  server.close(resolve);
-                })
-            ),
-          Promise.resolve()
-        )
-        .then(() => {
-          dummyServers = [];
-        });
-    });
-
-    function createDummyServers(n) {
-      process.env.WEBPACK_DEV_SERVER_BASE_PORT = 60000;
-
-      return (Array.isArray(n) ? n : [...new Array(n)]).reduce(
-        (p, _, i) =>
-          p.then(
-            () =>
-              new Promise((resolve, reject) => {
-                const server = http.createServer();
-
-                dummyServers.push(server);
-
-                server.listen(60000 + i, "0.0.0.0", () => {
-                  resolve();
-                });
-
-                server.on("error", (error) => {
-                  reject(error);
-                });
-              })
-          ),
-        Promise.resolve()
-      );
-    }
-
-    it("should returns the port when the port is specified", async () => {
-      const retryCount = 1;
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = retryCount;
-
-      const freePort = await Server.getFreePort(8082);
-
-      expect(freePort).toEqual(8082);
-    });
-
-    it("should returns the port when the port is null", async () => {
-      const retryCount = 2;
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = retryCount;
-
-      await createDummyServers(retryCount);
-
-      const freePort = await Server.getFreePort(null);
-
-      expect(freePort).toEqual(60000 + retryCount);
-    });
-
-    it("should returns the port when the port is undefined", async () => {
-      const retryCount = 3;
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = retryCount;
-
-      await createDummyServers(retryCount);
-
-      // eslint-disable-next-line no-undefined
-      const freePort = await Server.getFreePort(undefined);
-
-      expect(freePort).toEqual(60000 + retryCount);
-    });
-
-    it("should retry finding the port for up to defaultPortRetry times (number)", async () => {
-      const retryCount = 4;
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = retryCount;
-
-      await createDummyServers(retryCount);
-
-      const freePort = await Server.getFreePort();
-
-      expect(freePort).toEqual(60000 + retryCount);
-    });
-
-    it("should retry finding the port for up to defaultPortRetry times (string)", async () => {
-      const retryCount = 5;
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = retryCount;
-
-      await createDummyServers(retryCount);
-
-      const freePort = await Server.getFreePort();
-
-      expect(freePort).toEqual(60000 + retryCount);
-    });
-
-    it("should retry finding the port when serial ports are busy", async () => {
-      const busyPorts = [60000, 60001, 60002, 60003, 60004, 60005];
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = 1000;
-
-      await createDummyServers(busyPorts);
-
-      const freePort = await Server.getFreePort();
-
-      expect(freePort).toBeGreaterThan(60005);
-    });
-
-    it("should throws the error when the port isn't found", async () => {
-      expect.assertions(1);
-
-      jest.mock("portfinder", () => {
-        return {
-          getPortPromise: () => Promise.reject(new Error("busy")),
-        };
-      });
-
-      process.env.WEBPACK_DEV_SERVER_PORT_RETRY = 1;
-
-      try {
-        await Server.getFreePort();
-      } catch (error) {
-        expect(error.message).toMatchSnapshot();
-      }
     });
   });
 });
